Data Type

To simplify handling data for estimating the SD model, the package implements the DataSD struct as a subtype of Data. This is a mutable struct type that gathers the different data for search sessions. For model estimation, data from other types (e.g., DataFrame) needs to be gathered into this type. The Tutorials provide some examples of how this can be appropriately constructed.

Most fields are required, but some fill in nothing by default. For example, the search order or where consumers stopped scrolling may not be observed, in which case, these fields default to nothing.

StructuralSearchModels.DataSD — Type

Data type for the core Search and Discovery model. Indexing is based on sessions. See the tutorials for examples on how such data can be simulated or constructed.

Fields:

consumer_ids::Vector{Int}: consumer id for each session.

product_ids::Vector{Vector{Int}}: vector of vectors of product ids for each session.

product_characteristics::Vector{Matrix{T}}: product characteristics matrix.

positions::Vector{Vector{Int}}: positions for each session.

consideration_sets::Vector{Vector{Bool}}: consideration sets for each session, booleans whether searched or not.

purchase_indices::Vector{Int}: which product within session is purchased.

min_discover_indices::Union{Vector{Int}, Nothing}: index of last product must have been discovered in session (lowest position at which click occurred). This is mainly used during estimation, and can be constructed in with the fill_indices_min_discover! function.

stop_indices::Union{Vector{Int}, Nothing}: index of last product discovered in session (at which discovery stops). Is nothing if not available (e.g., when scrolling is not observed).

session_characteristics::Union{Vector{Vector{T}}, Nothing}: session characteristics for each session. Is nothing if not available (default).

search_paths::Union{Vector{Vector{Int}}, Nothing}: search paths for each session. Is nothing if search order is not available.

The data tracks everything on a session level, with the consumer_ids field allowing to track consumers over multiple sessions (which is currently not yet used).

Data can be accessed with d.fieldname, where d is a DataSD object. Indexing is then based on the different fields. For example, the following accesses the product characteristics of the first session in the data: d.product_characteristics[1].

For product characteristics, note that the last column is a dummy indicator for the outside option.

Convenience Functions

The package exports some convenience functions for this type:

d == d1 checks whether two DataSD objects are the same.

d[1:5] selects the first five search sessions from the data

fill_indices_min_discover!(data::DataSD) adds the min_discover_indices from the consideration sets.

It also offers a way to update the positions.

StructuralSearchModels.update_positions! — Function

update_positions!(data::DataSD, nA0, nd)

Update positions in data to fit the number of alternatives in initial awareness set nA0 and the number of alternatives per position nd.

Data Generation

Data from a search model model can be generated with the generate_data function. There are two versions.

The first one generates data from scratch, which requires specifying how many consumers and sessions to simulate (each consumer can have multiple sessions, which is not yet used). By default, it generates generic products using generate_products.

The second one takes products, sessions etc. as given and only simulates new search paths from model.

StructuralSearchModels.generate_data — Function

generate_data(model::Model, n_consumers, n_sessions_per_consumer;
                    n_A0 = 1, n_d = 1,
                    indices_list_characteristics = 1:length(m.β),
                    products = generate_products(n_consumers * n_sessions_per_consumer, MvNormal(I(length(m.β)-1))),
                    drop_undiscovered_products = false,
                    kwargs_path_generation...)

Generate and return data generated for the model model for n_consumers and n_sessions_per_consumer. By default, this assumes that there is one alternative in the initial awareness set (n_A0=1), one alternative per position (n_d=1), and to generate generic products using generate_products. Undiscovered products by default are not dropped. kwargs_path_generation are passed to the function generating the search paths.

generate_data(model::Model, data::Data;
                products = generate_products(data::Data),
                kwargs_path_generation...)

Generate and return data for the model model using the existing data object data. This allows simulating new search paths for the same consumers and products. If undiscovered products have been dropped, this function samples new products from data using generate_products. kwargs_path_generation are passed to the function generating the search paths.

StructuralSearchModels.generate_products — Function

generate_products(n_sessions, distribution::Distribution;
    n_products = 1_000_000,
    n_products_per_session = 30,
    outside_option = true,
    kwargs...)

Generate product IDs and their characteristics for n_sessions sessions. Each session has n_products_per_session products randomly sampled from n_products available in total. By default, n_products is large and 30 products are sampled per session.

Product characteristics for each product are sampled from the specified distribution, which needs to be multivariate (e.g., MvNormal) when using multiple characteristics. If outside_option=true, then an outside option is included as the first product for each session with product ID 0. For this product, the last characteristic is set to 1.0 and the remaining product characteristics are set to 0.0.

Returns product IDs and product characteristics as two separate vectors. Each element is a session, with a vector of product IDs or matrix of characteristics for all products available to the session.

Keyword options

The generate_data function passes kwargs_path_generation into the path generation. The following options currently are available:

seed: use a specific seed to generate data.

conditional_on_search=false: whether generate only search paths with at least one search. false by default.

conditional_on_search_iter=100: to generate search paths with at least one search for a particular session (with products etc.), the code tries by default up to 100 times using new draws for the shocks.