Estimation

For model estimation, this package implements estimators as subtypes of an abstract Estimator type. The Models section shows for each model which estimator is available. Currently, only simulated maximum likelihood is available, but I hope to implement more estimators (e.g., indirect inference) in the future.

Simulated Maximum Likelihood

The baseline Simulated Maximum Likelihood Estimator can be created with e = SMLE(n_draws) where n_draws specifies the number of simulation draws used for numerical integration.

StructuralSearchModels.SMLE — Type

SMLE(n_draws::Int; kwargs...)
SMLE(; numerical_integration_method, kwargs...)

Simulated maximum likelihood estimator (SMLE). This estimator is used to estimate models using maximum likelihood estimation when the likelihood function is not available in closed form.

The number of simulation draws must be specified either:

Via positional argument: SMLE(n_draws) creates an estimator with DefaultNI(n_draws).

Via keyword argument: SMLE(; numerical_integration_method = DefaultNI(100)) or any other NIMethod.

Arguments

n_draws::Int: (Positional constructor only) Number of simulation draws, creates DefaultNI(n_draws) as the integration method.

Optional Keyword Arguments

numerical_integration_method::NIMethod: (Required if not using positional constructor) Numerical integration method for main integration over shocks.

options_optimization::NamedTuple: Options for the optimization algorithm passed to OptimizationFunction of Optimization.jl. Uses LBFGS by default.

options_problem::NamedTuple: Options for the optimization problem passed to OptimizationProblem in Optimization.jl.

options_solver::NamedTuple: Options for the solver passed to solve in Optimization.jl. Defaults to 100,000 iterations.

numerical_integration_method_heterogeneity::NIMethod: Integration method for unobserved heterogeneity. Defaults to QMC(n_draws = 30).

conditional_on_search::Bool: Whether the likelihood function is conditional on search. Defaults to false.

Examples

# Using positional argument (most common)
e = SMLE(100)  # Creates SMLE with DefaultNI(100)

# With additional options
e = SMLE(200; conditional_on_search = true)

# Custom integration method
e = SMLE(; numerical_integration_method = QMC(n_draws = 500))

Options

Various options for the optimization and numerical integration procedures used in estimation can be specified. Moreover, the conditional_on_search option allows specifying whether the sample is selected based on there being at least one click.

Optimization

Simulated maximum likelhood finds the parameter estimates by numerically maximizing the model likelihood function. This is implemented using the amazing Optimization.jl package, which allows to specify a range of optimizers (from other packages) and solvers. The SMLE struct is set up to pass the respective options to the optimization package as keyword arguments.

Specifically, this is how the different fields are passed into the respective constructors and functions of the Optimization.jl package:

f = OptimizationFunction(loglikelihood, options_optimization.differentiation)
p = OptimizationProblem(f, startvals; options_problem...)
r = solve(p, options_optimization.algorithm; options_solver...)

Numerical Integration

Simulated maximum likelihood uses numerical integration to compute a likelihood function that has no closed-form. SMLE uses the DefaultNI, which is also the only option so far. For now, the DefaultNI uses the method introduced in Greminger (2025) to integrate over the unobserved shocks for all available models. This approach has the advantage that it does not require smoothing parameters (as KSFS) and is quite fast, but does not use the search order (as the GHK simulator does). In the future, more numerical integration methods, such as the GHK simulator that additionally uses the search order, should be implemented and made available as new types. The DefaultNI then should default to the appropriate method for each model and available data.

StructuralSearchModels.DefaultNI — Type

Default numerical integration that is used to integrate over the idiosyncratic shocks.

Options

n_draws::Int: Number of draws for main integration over shocks.

Estimation of Shock Variances

Options whether to estimate the variances of the unobserved shocks are directly passed as keyword arguments to the estimate function, rather than being part of the Estimator type. This is because it is not specific to any estimator and a rather general option.

By default, the variances are not estimated. To estimate them, pass the keyword argument estimation_shock_variances to the estimate function. The following are valid options:

estimation_shock_variances = [:σ_dE]: estimates the variance of the unobserved shock dE

estimation_shock_variances = [:σ_dV]: estimates the variance of the unobserved shock dV.

estimation_shock_variances = [:dUequaldE, :σ_dE]: estimates the variance of the unobserved shock dE and assumes that the distribution of the outside option shock dU0 is equal to that of dE.

estimation_shock_variances = [:dUequaldV, :σ_dV]: estimates the variance of the unobserved shock dV and assumes that the distribution of the outside option shock dU0 is equal to that of dV.

Note that the code currently does not check for the validity of the options and will try to estimate them as specified, even if it is not possible. For example, setting estimation_shock_variances = [:σ_dE, :σ_dV] will run the estimation and try to estimate both variances, even though they are not separately identified.

Fixing Parameters

By passing a fixed_parameters argument to the estimate function, parameters can be fixed so that they are not estimated. This is useful to evaluate identification of different parameters. For example, estimating the model as follows will fix the parameter β to the values provided in the model m:

estimate(m, SMLE(100), d; fixed_parameters = [:β])

Estimation Functions

For model estimation, the following abstract functions are available. These functions should be implemented for all available (and sensible) combinations of Model, Estimator, and Data types.

StructuralSearchModels.calculate_likelihood — Function

calculate_likelihood(model::Model, estimator::MLE, data::Data; kwargs...)

Calculate the likelihood of the model model given the data data and the estimator estimator.

Distributions.estimate — Function

estimate(model::Model, estimator::Estimator, data::Data; kwargs...)

Estimate the model using data and estimator. Returns the estimated model, the parameter estimates, the objective function at the estimates, the solver results, and the standard errors as tuple.

StructuralSearchModels.calculate_standard_errors — Function

calculate_standard_errors(model::Model, estimator::Estimator, data::Data; kwargs...)

Calculate standard errors for the model model using the specified data and estimator.