Welfare and revenue calculations

An important use case for structural search models is to compute revenue and consumer welfare effects of different models.

The following functions provide this functionality.

StructuralSearchModels.calculate_welfare — Function

calculate_welfare(model::Model, data::Data, n_sim; kwargs...)

Calculate consumer welfare for the model model using the data data and n_sim simulations.

StructuralSearchModels.calculate_revenues — Function

calculate_revenues(model::Model, data::Data,, kprice, n_draws; kwargs...)

Calculate revenues for the model model using the data data and n_draws simulation draws. kprice indicates which column in the characteristics table is used as price.

Model fit evaluation

To evaluate model fit, the following function can be used to compare moments in the data with those implied by the model.

StructuralSearchModels.calculate_fit_measures — Function

calculate_fit_measures(model::Model, data::Data, n_sim; kwargs...)

Compute fit measures of model for data. Allows for kwargs to be passed to the data generation.